+++ desktop-widgets/tab-widgets/TabBase.h
@@ -14,6 +14,8 @@ public:
 	virtual void updateData() = 0;
 	virtual void clear() = 0;
 	virtual void updateUi(QString titleColor);
+	virtual void enterEditMode();
+	virtual void exitEditMode();
 };
 
 #endif
