+++ desktop-widgets/tab-widgets/maintab.h
@@ -11,13 +11,8 @@
 #include <QTabWidget>
 
 #include "ui_maintab.h"
-#include "qt-models/completionmodels.h"
-#include "qt-models/divelocationmodel.h"
 #include "core/dive.h"
 #include "core/subsurface-qt/divelistnotifier.h"
-
-class ExtraDataModel;
-class DivePictureModel;
 
 class TabBase;
 class MainTab : public QTabWidget {
@@ -30,42 +25,15 @@ public:
 
 public
 slots:
-	void divesChanged(const QVector<dive *> &dives, DiveField field);
-	void diveSiteEdited(dive_site *ds, int field);
-	void tripChanged(dive_trip *trip, TripField field);
 	void updateDiveInfo();
-	void updateNotes(const struct dive *d);
-	void updateDateTime(const struct dive *d);
-	void updateTripDate(const struct dive_trip *t);
-	void updateDiveSite(struct dive *d);
-	void on_location_diveSiteSelected();
-	void on_locationPopupButton_clicked();
-	void on_diveguide_editingFinished();
-	void on_buddy_editingFinished();
-	void on_diveTripLocation_editingFinished();
-	void on_notes_editingFinished();
-	void on_duration_editingFinished();
-	void on_depth_editingFinished();
-	void on_dateEdit_editingFinished();
-	void on_timeEdit_editingFinished();
-	void on_rating_valueChanged(int value);
-	void on_tagWidget_editingFinished();
-	void closeWarning();
-	void escDetected(void);
-	void updateDateTimeFields();
+	void escDetected();
 	void colorsChanged();
 private:
 	Ui::MainTab ui;
-	bool ignoreInput; // When computionally editing fields, we have to ignore changed-signals
-	BuddyCompletionModel buddyModel;
-	DiveGuideCompletionModel diveGuideModel;
-	TagCompletionModel tagModel;
 	bool lastSelectedDive;
 	int lastTabSelectedDive;
 	int lastTabSelectedDiveTrip;
-	dive_trip *currentTrip;
 	QList<TabBase*> extraWidgets;
-	void divesEdited(int num); // Opens a warning window if more than one dive was edited
 	void changeEvent(QEvent *ev) override;
 	bool isDark;
 };
